correction des prix sans virgule et changement du nom de la bd

+++ database/migrations/2024_06_01_165715_create_produits_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('produits', function (Blueprint $table) {
             $table->id();
             $table->string('Nom');
-            $table->decimal('PrixDAchat', 10, 2); 
-            $table->decimal('PrixDVente', 10, 2); 
+            $table->integer('PrixDAchat'); // Colonne pour les prix d'achat sans virgule
+            $table->integer('PrixDVente'); // Colonne pour les prix de vente sans virgule
             $table->text('Alimentation')->nullable();
             $table->text('Description')->nullable();
             $table->string('Code_Barre_ean_13')->unique();
